Set autocmds as buffer local
The NeoVim "fidget" plugin creates an unnamed buffer and sets its filetype to "fidget".
However since this usually happens *after* this plugin was loaded, the setting of the filetype calls 'SetupAutoCommands()'.
The first thing it does it clear all the autocmds in the 'better_whitespace' augroup, but since the 'buftype' is now set to 'nofile' it will not re-add them.
Since the autocmds where set globally, the clearing of the autocmds cleared the ones from the previous buffer as well.

Steps to reproduce:
```
echo '   ' > file
nvim file --clean \
    -c 'lua vim.g.strip_whitespace_on_save=1' \
    -c 'source /path/to/plugin/vim-better-whitespace/plugin/better-whitespace.vim' \
    -c 'set filetype=' \ # To trigger the highlighting etc \
    -c 'lua bufid = vim.api.nvim_create_buf(false, true)' \
    -c 'lua vim.api.nvim_buf_set_option(bufid, "filetype", "fidget")' \
    -c ':w'
```

Without this fix the whitespaces will not be removed